Remove most font-awesome references in CSS
Most of the references were fontAwesome SVG icons that we'd hacked
directly in to the CSS for some reason I forget. No longer needed.

`.notification__message .fa` just becomes `.notification-group__icon
.icon`; it's for changing the color of the little "the following message in
your notification is a Follow notification" icon to be an on-theme color.
Changed the target so it still works.

Got rid of some styling for a haml template that doesn't exist anymore.

Notably I'm leaving in some `.fa` references because they are still in
Mastodon! Probably a bug I'll file with them.

https://github.com/mastodon/mastodon/blob/v4.5.1/app/javascript/styles/mastodon/widgets.scss#L50-L65